requirements.txt
boto3>=1.7.65
psycopg2-binary>=2.7.5
